+++ src/app/data/poe-ninja/service/currency-overview-http.service.ts
@@ -2,13 +2,18 @@ import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { environment } from '@env/environment';
 import { Observable } from 'rxjs';
-import { retry } from 'rxjs/operators';
+import { map, retry } from 'rxjs/operators';
 import { CurrencyOverviewResponse } from '../schema/currency-overview';
 
 export enum CurrencyOverviewType {
     Currency = 'Currency',
     Fragment = 'Fragment'
 }
+
+const PATH_TYPE_MAP = {
+    [CurrencyOverviewType.Currency]: 'currency',
+    [CurrencyOverviewType.Fragment]: 'fragments',
+};
 
 @Injectable({
     providedIn: 'root'
@@ -17,7 +22,7 @@ export class CurrencyOverviewHttpService {
     private readonly apiUrl: string;
 
     constructor(private readonly httpClient: HttpClient) {
-        this.apiUrl = `${environment.poeNinja.baseUrl}/data/currencyoverview`;
+        this.apiUrl = `${environment.poeNinja.baseUrl}/api/data/currencyoverview`;
     }
 
     public get(leagueId: string, type: CurrencyOverviewType): Observable<CurrencyOverviewResponse> {
@@ -30,7 +35,13 @@ export class CurrencyOverviewHttpService {
         return this.httpClient.get<CurrencyOverviewResponse>(this.apiUrl, {
             params
         }).pipe(
-            retry(3)
+            retry(3),
+            map(result => {
+                return {
+                    ...result,
+                    url: `${environment.poeNinja.baseUrl}/challenge/${PATH_TYPE_MAP[type]}`
+                }
+            })
         );
     }
 }