+++ src/app/layout/type/user-settings.type.ts
@@ -5,4 +5,5 @@ export interface UserSettings {
     language?: Language;
     openUserSettingsKeybinding?: string;
     exitAppKeybinding?: string;
+    displayVersion?: boolean;
 }