+++ src/app/modules/evaluate/component/evaluate-dialog/evaluate-dialog.component.spec.ts
@@ -63,7 +63,7 @@ describe('EvaluateDialogComponent', () => {
         })
             .compileComponents();
 
-        contextService = TestBed.get<ContextService>(ContextService);
+        contextService = TestBed.inject<ContextService>(ContextService);
         contextService.init({
             language: Language.English
         });